requirements.txt
django~=4.1.4
django-rest-framework
djangorestframework~=3.14.0
xmltodict~=0.13.0
django-crispy-forms~=1.14.0